chore: expose thoth in lib
Imports a patch previously exclusive to Botstopper.
